backend/requirements.txt
flask==3.0.2
numpy==1.26.4
opencv-python==4.9.0.80
scikit-image==0.22.0
scikit-learn
mtcnn==0.1.1
joblib==1.3.2
Werkzeug==3.0.1
python-dotenv==1.0.1
flask-cors==4.0.0
